+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - The `params` field of middlewares has been changed to the `any` type.
 - Enhance validation of `upstream` and `service` in the config package.
 - Remove `skip_test` from resolver.
+- Require `root` permissions to run as daemon.
 
 ### **Test**